% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_spotify_original.R
\docType{data}
\name{spotify_original}
\alias{spotify_original}
\title{Spotify Original Dataset}
\format{
A dataframe with 1226 rows (users) and 286 columns, one corresponding to the index number
(first column, \code{X}) and the remaining 285 corresponding to different artists.
}
\source{
Item-item collaborative filtering with binary or unitary data, \url{https://medium.com/radon-dev/item-item-collaborative-filtering-with-binary-or-unary-data-e8f0b465b2c3}. Accessed: 02 October 2024
}
\usage{
data(spotify_original)
}
\description{
This dataset contains data on musical tastes in Spotify from the \code{last.fm} dataset. It informs
about which users have listened to which artists. Refer to the source for details.
}
\examples{
data(spotify_original)
head(spotify_original)

}
\keyword{datasets}
